@@ -1,154 +1,112 @@
 {
   "name": "draft-js-plugins",
-  "version": "0.0.0",
-  "private": true,
   "description": "A Plugin Architecture on top of Draft.JS",
+  "version": "0.0.0",
   "author": {
     "name": "Nik Graf",
     "email": "nik@nikgraf.com",
     "url": "http://www.nikgraf.com"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/nikgraf/draft-js-plugins.git"
+  "dependencies": {
+    "draft-js": "^0.11.0",
+    "eslint-plugin-import": "^2.25.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   },
-  "workspaces": [
-    "draft-js-alignment-plugin",
-    "draft-js-anchor-plugin",
-    "draft-js-buttons",
-    "draft-js-counter-plugin",
-    "draft-js-divider-plugin",
-    "draft-js-drag-n-drop-plugin",
-    "draft-js-drag-n-drop-upload-plugin",
-    "draft-js-emoji-plugin",
-    "draft-js-focus-plugin",
-    "draft-js-hashtag-plugin",
-    "draft-js-image-plugin",
-    "draft-js-inline-toolbar-plugin",
-    "draft-js-linkify-plugin",
-    "draft-js-mention-plugin",
-    "draft-js-plugins-editor",
-    "draft-js-plugins-utils",
-    "draft-js-resizeable-plugin",
-    "draft-js-side-toolbar-plugin",
-    "draft-js-static-toolbar-plugin",
-    "draft-js-sticker-plugin",
-    "draft-js-undo-plugin",
-    "draft-js-video-plugin"
-  ],
-  "main": "lib/index.js",
+  "devDependencies": {
+    "@babel/core": "7.16.0",
+    "@babel/plugin-proposal-class-properties": "7.16.0",
+    "@babel/preset-env": "7.16.4",
+    "@babel/preset-react": "7.16.0",
+    "@babel/preset-typescript": "7.16.0",
+    "@changesets/cli": "^2.18.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-node-resolve": "^13.0.6",
+    "@storybook/react": "^6.3.12",
+    "@testing-library/jest-dom": "^5.15.0",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.0.3",
+    "@typescript-eslint/eslint-plugin": "4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^27.3.1",
+    "babel-loader": "^8.2.3",
+    "babel-plugin-transform-rename-import": "^2.3.0",
+    "css-loader": "0.28.11",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-import-resolver-typescript": "2.5.0",
+    "eslint-plugin-jsx-a11y": "4.0.0",
+    "eslint-plugin-no-only-tests": "^2.6.0",
+    "eslint-plugin-react": "^7.27.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "husky": "^7.0.4",
+    "jest": "^27.3.1",
+    "linaria": "2.1.1",
+    "lint-staged": "^12.0.3",
+    "matched": "^5.0.1",
+    "multer": "^1.4.3",
+    "postcss-loader": "^6.2.0",
+    "prettier": "^2.4.1",
+    "prismjs": "^1.25.0",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.60.0",
+    "style-loader": "^0.23.1",
+    "ts-jest": "27.0.7",
+    "ts-node": "10.4.0",
+    "typescript": "4.5.2"
+  },
   "keywords": [
-    "editor",
-    "wysiwyg",
-    "draft",
-    "react",
-    "ux",
     "components",
+    "draft",
+    "editor",
+    "react",
+    "react-component",
+    "ux",
     "widget",
-    "react-component"
+    "wysiwyg"
   ],
-  "scripts": {
-    "test": "mocha --compilers @babel/register --require testHelper.js \"./*/src/**/__test__/*.js\"",
-    "test:watch": "npm test -- --watch",
-    "lint": "npm run lint:eslint && npm run lint:flow",
-    "lint:eslint": "eslint --rule 'mocha/no-exclusive-tests: 2' ./",
-    "lint:eslint:fix": "eslint --fix --rule 'mocha/no-exclusive-tests: 2' ./",
-    "lint:flow": "flow status",
-    "alex": "alex README.md",
-    "storybook": "start-storybook -p 6006 -s ./stories/public",
-    "build-storybook": "build-storybook"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
+  "license": "MIT",
   "lint-staged": {
-    "*.js": [
+    "*.(js|ts|tsx)": [
       "prettier --write",
       "git add",
       "eslint"
+    ],
+    "*.(ts|tsx)": [
+      "bash -c tsc --noEmit"
     ]
   },
+  "main": "lib/index.js",
   "prettier": {
     "singleQuote": true,
     "trailingComma": "es5"
   },
-  "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.5.5",
-    "@storybook/react": "^5.1.11",
-    "alex": "^4.0.1",
-    "animateplus": "^1.4.0",
-    "autoprefixer": "^6.7.6",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-transform-rename-import": "^2.3.0",
-    "chai": "^3.5.0",
-    "chai-enzyme": "^1.0.0-beta.1",
-    "cheerio": "^0.22.0",
-    "cpx": "^1.5.0",
-    "css-loader": "^0.26.2",
-    "css-modules-require-hook": "^4.0.2",
-    "dirty-chai": "^1.2.2",
-    "draft-js-buttons": "latest",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^6.2.2",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-mocha": "^4.8.0",
-    "eslint-plugin-react": "^6.10.0",
-    "estraverse-fb": "^1.3.1",
-    "express": "^4.15.0",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "flow-bin": "^0.69.0",
-    "husky": "^3.0.4",
-    "jsdom": "^11.3.0",
-    "linaria": "^1.4.0-alpha.1",
-    "lint-staged": "^9.2.3",
-    "matched": "^4.0.0",
-    "mocha": "^3.0.2",
-    "multer": "^1.3.0",
-    "postcss-loader": "^1.3.3",
-    "prettier": "^1.18.2",
-    "prismjs": "^1.5.1",
-    "react-addons-pure-render-mixin": "^15.5.2",
-    "react-addons-test-utils": "^15.5.1",
-    "react-router": "^3.0.2",
-    "react-static-webpack-plugin": "2.1.0",
-    "react-transform-catch-errors": "^1.0.2",
-    "react-transform-hmr": "^1.0.4",
-    "redbox-react": "^1.3.4",
-    "rimraf": "^2.6.1",
-    "rollup": "^1.20.3",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "sinon": "^1.17.7",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.13.2",
-    "superagent": "^3.5.0",
-    "webpack": "^2.2.1",
-    "webpack-dev-middleware": "^1.10.1",
-    "webpack-hot-middleware": "^2.17.1"
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/draft-js-plugins/draft-js-plugins.git"
   },
-  "license": "MIT",
-  "dependencies": {
-    "draft-js": "^0.11.0",
-    "emojione": "^2.2.7",
-    "eslint-plugin-import": "^2.2.0",
-    "linkify-it": "^2.0.3",
-    "loader-utils": "^1.0.2",
-    "prepend-http": "1.0.4",
-    "react": "^16.9.0",
-    "react-custom-scrollbars": "^4.1.1",
-    "react-dom": "^16.9.0",
-    "react-icons": "^2.2.3",
-    "tlds": "^1.197.0",
-    "to-style": "^1.3.3"
-  }
+  "scripts": {
+    "alex": "alex README.md",
+    "build-storybook": "build-storybook",
+    "lint": "yarn lint:eslint && yarn tsc",
+    "lint:eslint": "eslint --ext .js,.ts,.tsx ./",
+    "lint:eslint:fix": "eslint --fix ./",
+    "storybook": "start-storybook -p 6006 -s ./stories/public",
+    "test": "yarn build:dts && yarn test:jest",
+    "test:jest": "jest",
+    "test:watch": "yarn test -- --watch",
+    "build": "yarn build:dts && yarn workspaces run build",
+    "build:dts": "yarn workspace @draft-js-plugins/editor run build && yarn workspace @draft-js-plugins/utils run build && yarn workspace @draft-js-plugins/buttons run build",
+    "prerelease": "yarn build",
+    "release": "changeset publish",
+    "postinstall": "husky install"
+  },
+  "workspaces": [
+    "packages/*"
+  ]
 }
