@@ -1,37 +1,43 @@
 const path = require('path');
 
 const packages = [
-  'draft-js-plugins-editor',
-  'draft-js-hashtag-plugin',
-  'draft-js-linkify-plugin',
-  'draft-js-anchor-plugin',
-  'draft-js-mention-plugin',
-  'draft-js-sticker-plugin',
-  'draft-js-undo-plugin',
-  'draft-js-emoji-plugin',
-  'draft-js-plugins-utils',
-  'draft-js-counter-plugin',
-  'draft-js-drag-n-drop-plugin',
-  'draft-js-drag-n-drop-upload-plugin',
-  'draft-js-inline-toolbar-plugin',
-  'draft-js-static-toolbar-plugin',
-  'draft-js-side-toolbar-plugin',
-  'draft-js-focus-plugin',
-  'draft-js-alignment-plugin',
-  'draft-js-image-plugin',
-  'draft-js-resizeable-plugin',
-  'draft-js-buttons',
-  'draft-js-video-plugin',
-  'draft-js-divider-plugin',
+  '@draft-js-plugins/editor',
+  '@draft-js-plugins/hashtag',
+  '@draft-js-plugins/linkify',
+  '@draft-js-plugins/anchor',
+  '@draft-js-plugins/mention',
+  '@draft-js-plugins/sticker',
+  '@draft-js-plugins/undo',
+  '@draft-js-plugins/emoji',
+  '@draft-js-plugins/utils',
+  '@draft-js-plugins/counter',
+  '@draft-js-plugins/drag-n-drop',
+  '@draft-js-plugins/drag-n-drop-upload',
+  '@draft-js-plugins/inline-toolbar',
+  '@draft-js-plugins/static-toolbar',
+  '@draft-js-plugins/side-toolbar',
+  '@draft-js-plugins/focus',
+  '@draft-js-plugins/alignment',
+  '@draft-js-plugins/image',
+  '@draft-js-plugins/resizeable',
+  '@draft-js-plugins/buttons',
+  '@draft-js-plugins/video',
+  '@draft-js-plugins/divider',
 ];
 
 const packagesAliases = {};
-packages.forEach(name => {
-  packagesAliases[name] = path.join(__dirname, '..', name, 'src');
+packages.forEach((name) => {
+  const [, folderName] = name.split('/');
+  packagesAliases[name] = path.join(
+    __dirname,
+    '../packages',
+    folderName,
+    'src'
+  );
 });
 
-module.exports = {
-  module: {
+module.exports = async ({ config }) => {
+  config.module = {
     rules: [
       {
         test: /\.css$/,
@@ -55,26 +61,32 @@ module.exports = {
       },
 
       {
-        test: /\.js$/,
-        loader: 'linaria/loader',
-        options: {
-          sourceMap: true,
-        },
+        test: /\.(js|jsx|ts|tsx)?$/,
+        use: [
+          { loader: 'babel-loader' },
+          {
+            loader: 'linaria/loader',
+            options: {
+              sourceMap: true,
+            },
+          },
+        ],
       },
-
       {
         test: /\.(png|jpg|gif|ico)$/,
         use: [{ loader: 'file-loader', options: { name: '[name].[ext]' } }],
       },
     ],
-  },
+  };
 
-  resolve: {
+  config.resolve = {
     alias: {
       ...packagesAliases,
       react: path.join(__dirname, '..', 'node_modules', 'react'),
       'prop-types': path.join(__dirname, '..', 'node_modules', 'prop-types'),
-      lodash: path.join(__dirname, '..', 'node_modules', 'lodash'),
     },
-  },
+    extensions: ['.ts', '.tsx', '.js', '.json', '.mjs'],
+  };
+
+  return config;
 };
