@@ -1,9 +1,18 @@
 import path from 'path';
-import nodeResolve from 'rollup-plugin-node-resolve';
-import babel from 'rollup-plugin-babel';
+import nodeResolve from '@rollup/plugin-node-resolve';
+import babel from '@rollup/plugin-babel';
+import { existsSync } from 'fs';
 
-const input = './src/index.js';
-const external = id => !id.startsWith('.') && !path.isAbsolute(id);
+const input = existsSync('./src/index.ts')
+  ? './src/index.ts'
+  : './src/index.tsx';
+const external = (id) => !id.startsWith('.') && !path.isAbsolute(id);
+const extensions = ['.ts', '.js', '.tsx', '.jsx'];
+const babelOptions = {
+  rootMode: 'upward',
+  extensions,
+  babelHelpers: 'bundled',
+};
 
 export default [
   {
@@ -14,7 +23,7 @@ export default [
       exports: 'named',
     },
     external,
-    plugins: [nodeResolve(), babel({ rootMode: 'upward' })],
+    plugins: [nodeResolve({ extensions }), babel(babelOptions)],
   },
   {
     input,
@@ -24,14 +33,19 @@ export default [
     },
     external,
     plugins: [
-      nodeResolve(),
+      nodeResolve({ extensions }),
       babel({
-        rootMode: 'upward',
+        ...babelOptions,
         plugins: [
           [
             'babel-plugin-transform-rename-import',
             {
-              replacements: [{ original: 'lodash', replacement: 'lodash-es' }],
+              replacements: [
+                {
+                  original: 'lodash',
+                  replacement: 'lodash-es',
+                },
+              ],
             },
           ],
         ],
